Add automatic Public Suffix List updates
This takes care of creating the file outright if it does not yet exist,
and keeping it up-to-date on a rolling basis. This means you no longer
need a cronjob to keep it up-to-date. It also adds a sanity test to
confirm that the list is functional before using it.